Refactor onDateRangeChange
